♻️ Refactor: application-local.yml 삭제, PageDTO, BaseEntity 위치 변경, @EnableJpaAuditing 추가
