+++ src/main/java/com/umc9th/bizscan/domain/aiAnalysis/entity/Swot.java
@@ -1,6 +1,6 @@
 package com.umc9th.bizscan.domain.aiAnalysis.entity;
 
-import com.umc9th.bizscan.global.common.BaseEntity;
+import com.umc9th.bizscan.global.entity.BaseEntity;
 import jakarta.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
