+++ src/main/java/com/umc9th/bizscan/domain/store/entity/Tag.java
@@ -1,6 +1,6 @@
 package com.umc9th.bizscan.domain.store.entity;
 
-import com.umc9th.bizscan.global.common.BaseEntity;
+import com.umc9th.bizscan.global.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
