--- src/main/java/com/umc9th/bizscan/global/common/JpaAuditingConfig.java
@@ -1,8 +1,0 @@
-package com.umc9th.bizscan.global.common;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@Configuration
-public class JpaAuditingConfig {}
