+++ src/main/resources/application-dev.yml
@@ -1,1 +1,1 @@
-spring.application.name=bizscan
+
