--- src/main/resources/application-local.yml
@@ -1,1 +1,0 @@
-spring.application.name=bizscan
